pyproject.toml
[project]
name = "ghostops"
version = "0.1.0"
description = "Command and Control Framework."
authors = [
    {name = "Awagat Dhungana",email = "4w4647@gmail.com"}
]
license = {text = "BSD 3-Clause"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama (>=0.4.6,<0.5.0)",
    "keystone-engine (>=0.9.2,<0.10.0)",
    "tabulate (>=0.9.0,<0.10.0)"
]

[project.scripts]
ghostops = 'ghostops.ghostops:main'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"